requirements.txt
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
iniconfig==1.1.1
packaging==20.9
Pillow==8.1.2
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
requests==2.25.1
soupsieve==2.2
toml==0.10.2
urllib3==1.26.3
